Trying to make sense of the advice I think it means that for all calls except for read, it is important that the char to screen redirection is faithful. This makes sense as for ndread it is not harmful to misreport one char are not present (this will usually result in another checking call). Similarly for flushing buffers, we cannot "overflush" only "underflush" thus using the int interface is fine.
For read, I now turn off interrupts after verifying that the screen is fine and enter a simulated interrupt call to the original BIOS handler to get the char. Once we are back and the char has been placed in the buffer we turn off interrupts and check the screen number again.
This makes more sense than calling the simulated int 36h routine which inevitably ends up doing the same thing but would result in blocking the thread on pScrIoOk or bKeybWait as opposed to the more sensible (for when the request screen number is known) currentSIB.